requirements.txt
torch
Pillow
fastapi
uvicorn
opencv-python-headless
websockets
numpy
timm
gitpython
setuptools
pandas